advanced to 6-1

+++ app/controllers/staff/programs_controller.rb
@@ -42,8 +42,20 @@ class Staff::ProgramsController < Staff::Base
 
   def destroy
     program = Program.find(params[:id])
-    program.destroy!
-    flash.notice = 'プログラムを削除しました。'
+    if program.deletable?
+      program.destroy!
+      flash.notice = 'プログラムを削除しました。'
+    else
+      flash.alert = 'このプログラムは申し込みがあるため、削除できません。'
+    end
+    redirect_to :staff_programs
+  end
+
+  def entries
+    entries_form = Staff::EntriesForm.new(Program.find(params[:id]))
+    entries_form.assign_attributes(params[:form])
+    entries_form.save
+    flash.notice = 'プログラム申し込みのフラグを更新しました。'
     redirect_to :staff_programs
   end
 end
